More Methodized Queries
AppMan_NewAccount finished (I think).
Also put log queries into the QueryManager.
(The way logs are done is really weird. You only need ONE query for
every single log, but the way Floris did it is that there are many
different types of queries for each logs... This is super inefficient
and should be fixed.)